Implemented plotting via sequencer. @todo Reduce rendering per point.

+++ Visible/include/imGuiCustom/ImSequencer.h
@@ -22,11 +22,11 @@ namespace ImSequencer
         bool focused = false;
         virtual int GetFrameMin() const = 0;
         virtual int GetFrameMax() const = 0;
-        virtual int GetItemCount() const = 0;
+        virtual size_t GetItemCount() const = 0;
 
         virtual void BeginEdit(int /*index*/) {}
         virtual void EndEdit() {}
-        virtual int GetItemTypeCount() const { return 0; }
+        virtual size_t GetItemTypeCount() const { return 0; }
         virtual const char *GetItemTypeName(int /*typeIndex*/) const { return ""; }
         virtual const char *GetItemLabel(int /*index*/) const { return ""; }
